Moved column filtering responsibility from preprocess.sh to data_wrangling.R for readability & modifiability; improved index creation accuracy

+++ jobscript.sh
@@ -142,10 +142,12 @@ else
     fi
 fi
 
-# Delete Shiny-specific data
+# Delete Shiny-specific data unless -s flag is used
 
-rm ../clean/data.rds
-
+if ! $skip_shiny_app; then
+echo "Removing shiny-specific data frame..."
+    rm ../clean/data.rds
+fi
 
 # All scripts run
